Extract lifecycle handling and share for Kafka and Pulsar readers
