+++ graylog2-server/src/main/java/org/graylog2/shared/messageq/pulsar/PulsarMessageQueueReader.java
@@ -21,7 +21,8 @@ import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
 import com.google.common.collect.ImmutableList;
-import com.google.common.util.concurrent.AbstractExecutionThreadService;
+import com.google.common.eventbus.EventBus;
+import com.google.common.util.concurrent.Uninterruptibles;
 import org.apache.pulsar.client.api.BatchReceivePolicy;
 import org.apache.pulsar.client.api.ConsumerInterceptor;
 import org.apache.pulsar.client.api.Message;
@@ -32,9 +33,9 @@ import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.api.Schema;
 import org.graylog2.plugin.journal.RawMessage;
 import org.graylog2.shared.buffers.ProcessBuffer;
+import org.graylog2.shared.messageq.AbstractMessageQueueReader;
 import org.graylog2.shared.messageq.MessageQueue;
 import org.graylog2.shared.messageq.MessageQueueException;
-import org.graylog2.shared.messageq.MessageQueueReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,10 +47,10 @@ import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 
 import static com.codahale.metrics.MetricRegistry.name;
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 @Singleton
-public class PulsarMessageQueueReader extends AbstractExecutionThreadService implements MessageQueueReader {
-
+public class PulsarMessageQueueReader extends AbstractMessageQueueReader {
     private static final Logger LOG = LoggerFactory.getLogger(PulsarMessageQueueReader.class);
 
     private final String name;
@@ -69,7 +70,11 @@ public class PulsarMessageQueueReader extends AbstractExecutionThreadService imp
 
     @Inject
     public PulsarMessageQueueReader(MetricRegistry metricRegistry,
-                                    Provider<ProcessBuffer> processBufferProvider) {
+            Provider<ProcessBuffer> processBufferProvider,
+            EventBus eventBus) {
+
+        super(eventBus);
+
         // Using a ProcessBuffer directly will lead to guice error:
         // "Please wait until after injection has completed to use this object."
         this.processBufferProvider = processBufferProvider;
@@ -85,6 +90,8 @@ public class PulsarMessageQueueReader extends AbstractExecutionThreadService imp
 
     @Override
     protected void startUp() throws Exception {
+        super.startUp();
+
         LOG.info("Starting pulsar message queue reader service: {}", name);
 
         this.client = PulsarClient.builder()
@@ -112,16 +119,21 @@ public class PulsarMessageQueueReader extends AbstractExecutionThreadService imp
             client.close();
             client.shutdown();
         }
+
+        super.shutDown();
     }
 
     @Override
     protected void run() throws Exception {
-        // TODO pause processing when LifeCycle changes
         // TODO add metrics
         // TODO the JournalReader limits the read based on the remaining capacity in the processBuffer
         //      do we need this?   final long remainingCapacity = processBuffer.getRemainingCapacity();
 
         while (isRunning()) {
+            if (!shouldBeReading()) {
+                Uninterruptibles.sleepUninterruptibly(100, MILLISECONDS);
+                continue;
+            }
             final List<MessageQueue.Entry> entries = read();
             entries.forEach(entry -> {
                 LOG.info("Consumed message: {}", entry);
